+++ index.ts
@@ -1,3 +1,3 @@
-export { Schema } from './lib/schema'
+export { Schema, SchemaOptions, SchemaProperties, SchemaValidationOptions } from './lib/schema'
 export { metadataFor } from './lib/metadata'
 export * from './lib/decorator'