+++ CHANGELOG.md
@@ -2,6 +2,63 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.2.4"></a>
+## [0.2.4](https://github.com/vitarn/tdv/compare/v0.2.3...v0.2.4) (2018-03-31)
+
+
+
+<a name="0.2.3"></a>
+## [0.2.3](https://github.com/vitarn/tdv/compare/v0.2.2...v0.2.3) (2018-03-31)
+
+
+### Bug Fixes
+
+* **schema:** build method skip parse props if not present ([0ecb228](https://github.com/vitarn/tdv/commit/0ecb228))
+
+
+
+<a name="0.2.2"></a>
+## [0.2.2](https://github.com/vitarn/tdv/compare/v0.2.1...v0.2.2) (2018-03-31)
+
+
+### Bug Fixes
+
+* **schema:** add static property Joi new instance ([8ec6375](https://github.com/vitarn/tdv/commit/8ec6375))
+
+
+
+<a name="0.2.1"></a>
+## [0.2.1](https://github.com/vitarn/tdv/compare/v0.2.0...v0.2.1) (2018-03-31)
+
+
+### Bug Fixes
+
+* **schema:** validate accept options extends joi option with apply field ([775ea5d](https://github.com/vitarn/tdv/commit/775ea5d))
+
+
+
+<a name="0.2.0"></a>
+# [0.2.0](https://github.com/vitarn/tdv/compare/v0.1.4...v0.2.0) (2018-03-31)
+
+
+### Features
+
+* **schema:** add Schema.build Schema#parse ([cfb7328](https://github.com/vitarn/tdv/commit/cfb7328))
+
+
+
+<a name="0.1.4"></a>
+## [0.1.4](https://github.com/vitarn/tdv/compare/v0.1.3...v0.1.4) (2018-03-31)
+
+
+### Bug Fixes
+
+* **decorator:** set joi label ([b814fd9](https://github.com/vitarn/tdv/commit/b814fd9))
+* **joi:** try require ([1903374](https://github.com/vitarn/tdv/commit/1903374))
+* **schema:** tdv:ref is typeof Schema ([dda14cf](https://github.com/vitarn/tdv/commit/dda14cf))
+
+
+
 <a name="0.1.3"></a>
 ## [0.1.3](https://github.com/vitarn/tdv/compare/v0.1.2...v0.1.3) (2018-03-29)